+++ src/main/resources/application.properties
@@ -1,1 +1,1 @@
-
+server.port=9090
